% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delventhal_model.R
\name{dummyfun_1}
\alias{dummyfun_1}
\title{Create \mjseqn{d_{1t}}}
\usage{
dummyfun_1(t, tau_2)
}
\arguments{
\item{t}{An \verb{integer vector} \mjseqn{1, \dots, T}}

\item{tau_2}{An \verb{integer scalar}}
}
\value{
Returns an integer vector
}
\description{
\loadmathjax
Creates the dummy variable \mjseqn{d_{1t} = 1\{t \leq \tau_2\}}
}
