% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delventhal_model.R
\name{dummyfun_2}
\alias{dummyfun_2}
\title{Create \mjseqn{d_{2t}}}
\usage{
dummyfun_2(t, tau_2, tau_3)
}
\arguments{
\item{t}{An \verb{integer vector} \mjseqn{1, \dots, T}}

\item{tau_2}{An \verb{integer scalar}}

\item{tau_3}{An \verb{integer scalar}}
}
\value{
Returns an integer vector
}
\description{
\loadmathjax
Creates the dummy variable \mjseqn{d_{2t} = 1\{\tau_2 < t < \tau_3\}}
}
