% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delventhal_model.R
\name{dummyfun_3}
\alias{dummyfun_3}
\title{Create \mjseqn{d_{3t}}}
\usage{
dummyfun_3(t, tau_3)
}
\arguments{
\item{t}{An \verb{integer vector} \mjseqn{1, \dots, T}}

\item{tau_3}{An \verb{integer scalar}}
}
\value{
Returns an integer vector
}
\description{
\loadmathjax
Creates the dummy variable \mjseqn{d_{3t} = 1\{t \geq \tau_3\}}
}
