% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delventhal_model.R
\name{fit_DT_model}
\alias{fit_DT_model}
\title{Fit the Delventhal model by OLS}
\usage{
fit_DT_model(tau_2, tau_3, y, x_t)
}
\arguments{
\item{tau_2}{tau_2 An \verb{integer scalar}}

\item{tau_3}{tau_2 An \verb{integer scalar}}

\item{y}{\verb{integer vector} The dependent variable, e.g. the crude birth or death rate}

\item{x_t}{\verb{integer vector} The regressor \mjseqn{x_t}, in the simplest case a vector of ones}
}
\value{
Returns a list of parameters and model diagnostics:
\itemize{
\item \code{alpha} Contains the parameters \mjseqn{\alpha_1} and\mjseqn{\alpha_3}
\item \code{sigma} Contains the parameters \mjseqn{\sigma^2_1} and\mjseqn{\\sigma^2_3}
\item \code{tau} Contains the parameters \mjseqn{\tau_1} and\mjseqn{\tau_3}
\item \code{y} The original dependent variable
\item \code{y_hat} The predicted values
\item \code{RSS} The residual sum of squares
}
}
\description{
Fit the Delventhal model by OLS
}
\examples{
T_max = 100
x_t = rep(1, T_max)
tau_2 = 33
tau_3 = 67
y = matrix(make_z1_scalar(1:T_max, tau_2, tau_3)*10 + rnorm(T_max, sd = 0.2),
          ncol = 1)
DT_fit = fit_DT_model(tau_2, tau_3, y, x_t)
}
