% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delventhal_model.R
\name{grid_Search_best_DT_fit}
\alias{grid_Search_best_DT_fit}
\title{Select the best fit by grid searchign over \mjseqn{(\tau_1, \tau_3)}}
\usage{
grid_Search_best_DT_fit(y, tau_3_min = 0)
}
\arguments{
\item{y}{\verb{integer vector} The dependent variable, e.g. the crude birth or death rate}

\item{tau_3_min}{\verb{integer scalar} Lower bound for \mjseqn{tau_3}}
}
\value{
Returns the same list as \code{\link{fit_DT_model}}
}
\description{
Select the best fit by grid searchign over \mjseqn{(\tau_1, \tau_3)}
}
\examples{
T_max = 100
x_t = rep(1, T_max)
tau_2 = 33
tau_3 = 67
y = matrix(make_z1_scalar(1:T_max, tau_2, tau_3)*10 + rnorm(T_max, sd = 0.2),
ncol = 1)
DT_fit = grid_Search_best_DT_fit(y)
}
