% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delventhal_model.R
\name{predict_DT_model}
\alias{predict_DT_model}
\title{Predict \mjseqn{y_t} based on estimated parameters}
\usage{
predict_DT_model(alpha, tau, y)
}
\arguments{
\item{alpha}{\verb{numeric vector} of length 2: \mjseqn{(\alpha_1, \alpha_3)}}

\item{tau}{\verb{numeric vector} of length 2: \mjseqn{(\tau_1, \tau_3)}}

\item{y}{\verb{numeric vector} with the same length as the original \mjesqn{y_t}}
}
\value{
\verb{numeric vector}
}
\description{
Predict \mjseqn{y_t} based on estimated parameters
}
\examples{
T_max = 100
x_t = rep(1, T_max)
tau_2 = 33
tau_3 = 67
y = matrix(make_z1_scalar(1:T_max, tau_2, tau_3)*10 + rnorm(T_max, sd = 0.2),
          ncol = 1)
DT_fit = fit_DT_model(tau_2, tau_3, y, x_t)

predict_DT_model(DT_fit$alpha,
                 DT_fit$tau,
                 DT_fit$y)
}
